Added explicit casts in error tests

+++ tests/errors/test_pow3.py
@@ -1,6 +1,8 @@
+from ltypes import i32, i64
+
 def f():
     x: list[i32]
-    x = [2,3]
+    x = [2, 3]
     a: i64
-    a = 3
-    print(pow(x,a))
+    a = i64(3)
+    print(pow(x, a))
